ml-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0
numpy==1.26.2
pandas==2.1.3
scikit-learn==1.3.2
tensorflow==2.15.0
prophet==1.1.5
xgboost==2.0.3
scipy==1.11.4
matplotlib==3.8.2
seaborn==0.13.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
redis==5.0.1
requests==2.31.0
python-multipart==0.0.6
